requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.64.0
diffusers>=0.25.0
transformers>=4.30.0
datasets>=2.10.0
huggingface_hub>=0.16.0
peft>=0.4.0
psutil>=5.9.0
xformers>=0.0.20